+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -5,6 +5,11 @@
 - Added `MessageListView.paginationLimit`
 - Allow the various ListView widgets to be themed via ThemeData classes
 
+🔄 Changed
+
+- `StreamChat.of(context).user` is now deprecated in favor of `StreamChat.of(context).currentUser`.
+- `StreamChat.of(context).userStream` is now deprecated in favor of `StreamChat.of(context).currentUserStream`.
+
 🐞 Fixed
 
 - Fix floating date divider not having a fixed size